requirements.txt
async-timeout==5.0.1
bleak==2.1.1
certifi==2026.1.4
charset-normalizer==3.4.4
dbus-fast==4.0.0
grpcio==1.78.0
grpcio-reflection==1.78.0
grpcio-tools==1.78.0
idna==3.11
meshtastic==2.7.7
numpy==2.2.6
packaging==24.2
pandas==2.3.3
protobuf==6.33.5
Pypubsub==4.0.7
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
